+++ src/qpe_toolbox/circuit/__init__.py
@@ -21,11 +21,6 @@ from .parametrized_circuits import (
     generate_rand_circuit,
 )
 from .plot_circuits import draw_layered_circuit, draw_layered_expval
-from .qaoa import (
-    brute_force_maxcut,
-    compute_qaoa_contraction_costs,
-    study_optimization_time_costs,
-)
 from .serialize_circuits import (
     deserialize_to_qiskit_QuantumCircuit,
     deserialize_to_quimb_Circuit,